requirements.txt
opencv-python
flask
azure-cognitiveservices-vision-face
azure.cosmos
requests
flask_cors
azure.cognitiveservices.speech